% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/src.R
\name{src}
\alias{src}
\alias{src_couchdb}
\alias{src_elastic}
\alias{src_solr}
\title{Setup a database connection}
\usage{
src_elastic(x, ...)

src_solr(x, ...)

src_couchdb(x, ...)
}
\arguments{
\item{x}{aasdfasdfsf}

\item{...}{adfafasfasdf}
}
\description{
Setup a database connection
}
\examples{
\dontrun{
src_elastic()
src_solr()
src_couchdb()
}
}

